newcloudproject/requirements.txt
streamlit==1.26.0
python-dotenv==1.0.1
groq==0.13.0 